Move FreeStack into its own module
