+++ ipc_test/Cargo.toml
@@ -17,3 +17,4 @@ page_size = "0.5.0"
 raw_sync = "0.1.5"
 sendfd = "0.4.3"
 serde = { version = "1.0.149", features = ["derive"] }
+log = "0.4.17"
