+++ ipc_test/src/lib.rs
@@ -1,10 +1,14 @@
+mod freestack;
+
 use std::{
     fs::File,
     os::fd::{AsRawFd, FromRawFd},
     slice::{from_raw_parts, from_raw_parts_mut},
-    sync::atomic::{AtomicU8, Ordering},
+    sync::atomic::{AtomicU8, Ordering}, backtrace::Backtrace,
 };
 
+use freestack::FreeStack;
+use log::trace;
 use memfd::{FileSeal, HugetlbSize, MemfdOptions};
 use memmap2::{MmapOptions, MmapRaw};
 
@@ -96,90 +100,6 @@ fn align_to(size: usize, alignment: usize) -> usize {
         alignment * (div + 1)
     } else {
         alignment * div
-    }
-}
-
-/// A stack as an array that is stored somewhere in (shared) memory
-struct FreeStack {
-    ///
-    /// Pointer to the underlying memory
-    ///
-    /// The layout is as follows:
-    ///
-    ///
-    /// +------------+--------+-----+--------+
-    /// | top: usize | elem 0 | ... | elem N |
-    /// +------------+--------+-----+--------+
-    ///
-    /// All elements are `usize`, too.
-    ///
-    base_ptr: *mut u8,
-
-    /// Total size of memory that we have available
-    size_in_bytes: usize,
-}
-
-impl FreeStack {
-    pub fn new(base_ptr: *mut u8, size_in_bytes: usize) -> Self {
-        FreeStack {
-            base_ptr,
-            size_in_bytes,
-        }
-    }
-
-    pub fn get_stack_idx(&self) -> usize {
-        let slice = unsafe { from_raw_parts(self.base_ptr, std::mem::size_of::<usize>()) };
-        *bytemuck::from_bytes(slice)
-    }
-
-    fn set_stack_idx(&mut self, idx: usize) {
-        let slice = unsafe { from_raw_parts_mut(self.base_ptr, std::mem::size_of::<usize>()) };
-        let slice_usize: &mut [usize] = bytemuck::cast_slice_mut(slice);
-        slice_usize[0] = idx;
-    }
-
-    fn get_elems(&self) -> &[usize] {
-        let slice = unsafe {
-            let elems_ptr = self
-                .base_ptr
-                .offset(std::mem::size_of::<usize>().try_into().unwrap());
-            from_raw_parts(elems_ptr, self.size_in_bytes - std::mem::size_of::<usize>())
-        };
-        bytemuck::cast_slice(slice)
-    }
-
-    fn get_elems_mut(&mut self) -> &mut [usize] {
-        let slice = unsafe {
-            let elems_ptr = self
-                .base_ptr
-                .offset(std::mem::size_of::<usize>().try_into().unwrap());
-            from_raw_parts_mut(elems_ptr, self.size_in_bytes - std::mem::size_of::<usize>())
-        };
-        bytemuck::cast_slice_mut(slice)
-    }
-
-    pub fn pop(&mut self) -> Option<usize> {
-        let idx = self.get_stack_idx();
-
-        if idx == 0 {
-            None
-        } else {
-            let elems = self.get_elems();
-            let elem = elems[idx - 1];
-
-            self.set_stack_idx(idx - 1);
-
-            Some(elem)
-        }
-    }
-
-    pub fn push(&mut self, new_elem: usize) {
-        let idx = self.get_stack_idx();
-        let elems = self.get_elems_mut();
-
-        elems[idx] = new_elem;
-
-        self.set_stack_idx(idx + 1);
     }
 }
 
@@ -470,7 +390,7 @@ impl SharedSlabAllocator {
 
 impl Drop for SharedSlabAllocator {
     fn drop(&mut self) {
-        // println!("SharedSlabAllocator::drop: {}", Backtrace::force_capture())
+        trace!("SharedSlabAllocator::drop: {}", Backtrace::force_capture())
     }
 }
 
@@ -478,9 +398,7 @@ impl Drop for SharedSlabAllocator {
 mod test {
     use std::time::Duration;
 
-    use memmap2::MmapOptions;
-
-    use crate::{align_to, FreeStack, SharedSlabAllocator, SlotForWriting};
+    use crate::{align_to, SharedSlabAllocator, SlotForWriting};
 
     #[test]
     fn test_align() {
@@ -488,55 +406,6 @@ mod test {
         assert_eq!(align_to(1, 4096), 4096);
         assert_eq!(align_to(4096, 4096), 4096);
         assert_eq!(align_to(4097, 4096), 2 * 4096);
-    }
-
-    #[test]
-    fn test_stack_happy() {
-        // raw memory length in bytes
-        let raw_length = 4096;
-
-        // get some "raw" memory
-        let mut mmap = MmapOptions::new()
-            .len(raw_length)
-            .map_anon()
-            .expect("could not mmap");
-
-        let base_ptr = mmap.as_mut_ptr();
-        let mut s = FreeStack::new(base_ptr, raw_length);
-
-        assert!(s.pop().is_none());
-
-        s.push(1);
-        s.push(2);
-        s.push(3);
-
-        assert_eq!(s.pop(), Some(3));
-        assert_eq!(s.pop(), Some(2));
-        assert_eq!(s.pop(), Some(1));
-        assert!(s.pop().is_none());
-    }
-
-    #[test]
-    #[should_panic(expected = "index out of bounds: the len is 3 but the index is 3")]
-    fn test_stack_unhappy() {
-        // raw memory length in bytes
-        let raw_length = std::mem::size_of::<usize>() * 4;
-
-        // get some "raw" memory
-        let mut mmap = MmapOptions::new()
-            .len(raw_length)
-            .map_anon()
-            .expect("could not mmap");
-
-        let base_ptr = mmap.as_mut_ptr();
-        let mut s = FreeStack::new(base_ptr, raw_length);
-
-        assert!(s.pop().is_none());
-
-        s.push(1);
-        s.push(2);
-        s.push(3);
-        s.push(4); // panic! out of bounds.
     }
 
     #[test]
